skillradar-ai/Skills-Mirage-System/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1
xgboost==2.0.3
sentence-transformers==2.7.0
pydantic==2.6.4
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-experimental>=0.0.1
duckduckgo-search>=4.0.0
python-dotenv>=1.0.0
ddgs>=9.11.2
